% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_dorp_brandstof.R
\name{boot_dorp_brandstof}
\alias{boot_dorp_brandstof}
\title{boot_dorp_brandstof}
\usage{
boot_dorp_brandstof(pop, dta, placename = "place",
  unitname = "sol_energy_coal_consumptionwinter",
  formatname = "sol_energy_coal_format", notusename = "none",
  unitweights, rep = 1000, plt = TRUE, debug = FALSE,
  dropperc = 0.025, graphdir = "Grafieke", stoor = TRUE)
}
\arguments{
\item{pop}{dataframe}

\item{dta}{data frame giving indication of proportions who use the fuel and format}

\item{placename}{Character. Default: "place"}

\item{unitname}{Character. Default: "sol_energy_coal_consumptionwinter"}

\item{formatname}{Character. Default: "sol_energy_coal_format"}

\item{notusename}{Character. Default: "none"}

\item{unitweights}{Named list with unit weights}

\item{rep}{Integer. Default: 1000}

\item{plt}{Logical. Plot or not. Default: TRUE}

\item{debug}{Logical.}

\item{dropperc}{Numeric. Drop proportions smaller than dropperc. Default 0.025}

\item{graphdir}{Character. Directory to write graphs to. Default: "Grafieke",}

\item{stoor}{Logical: Save output or not. Default: TRUE}
}
\description{
Function to estimate township fuel use with bootstrapping
}
