% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootformula.R
\name{bootformula}
\alias{bootformula}
\title{bootformula}
\usage{
bootformula(frm = "mean(rpois(n = 31, lambda = 0.73)) * 31", rep = 1000)
}
\arguments{
\item{frm}{Formula}

\item{rep}{Integer. Number of repeat}
}
\description{
Generate values from a formula using bootstrapping
}
