% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveGammaparams.R
\name{deriveGammaParams}
\alias{deriveGammaParams}
\title{deriveGammaParams}
\usage{
deriveGammaParams(Mean = NULL, Sd = NULL, Var = NULL)
}
\arguments{
\item{Mean}{Numeric. The mean}

\item{Sd}{Numeric. The standard deviation}

\item{Var}{Numeric. The variance (alternative to sd)}
}
\description{
Derive the parameters for the Gamma distrubution from the mean and sd
}
