% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genereerTydMat.R
\name{genereerTydMat}
\alias{genereerTydMat}
\title{genereerTydMat}
\usage{
genereerTydMat(week_p = c(6/7, 1/7), uur_p = c(23/24, 1/24))
}
\arguments{
\item{week_p}{Numeriese vektor van lengte 2 met waarskynlikehede van vir 0 en 1 (ja en). Verstek 1 dag per week}

\item{uur_p}{Numeriese vektor van lengte 24 met waarskynlikehede vir uur van dag. Verstek 1 uur per dag}
}
\value{
Matriks van 7x24
}
\description{
genereer 'n matriks met 0 en 1 vir gebruik in Aermod ewekansig met waarskynlikheid vir dag en uur
}
