% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_uur.R
\name{kg_uur}
\alias{kg_uur}
\title{Kg Hour}
\usage{
kg_uur(x = tt, propvar = "prop", fuelprop = 0.65, nnn = c(200, 14),
  mcspecfn = "~/tmp/MC_estimation_kg_coal_hod_MC_vars",
  mcspecfn.add = NULL, mcsimfn = "~/tmp/MC_estimation_kg_coal_hod",
  mcsimfn.add = NULL, yl = 1000)
}
\arguments{
\item{x}{Data frame such as the tt data frame generated in binconf.uur}

\item{propvar}{Character vector. Name of the proportion column}

\item{fuelprop}{Numeric. The fuel proportion}

\item{nnn}{Numeric vector. The number of timesteps for use in monte carlo simulation}

\item{mcspecfn}{Character Vector. Filename for MC simulation of specific variables}

\item{mcspecfn.add}{Character vector. Additional extention to filename mcspecfn}

\item{mcsimfn}{Character Vector. Filename for MC simulation of coal burned per hour of day}

\item{mcsimfn.add}{Character vector. Additional extention to filename mcsimfn}

\item{yl}{Numeric. Y axis limit}
}
\description{
Uses output of binconf.uur and obtains kg per hour while also plotting a graph using Monte Carlo
}
