% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maak_distlist.R
\name{maak_distlist}
\alias{maak_distlist}
\title{maak.distlist}
\usage{
maak_distlist(nm, dist, params)
}
\arguments{
\item{nm}{Character. Name of variable}

\item{dist}{Character. Name of distribution.}

\item{params}{Numeric. Named parameter values}
}
\description{
Create an instiance of the class distlist
}
