% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maakIndicresTab.R
\name{maakIndicTab}
\alias{maakIndicTab}
\alias{maakIndicResTab}
\title{maakIndicResTab}
\usage{
maakIndicResTab(
  reftab,
  i = indicators,
  props,
  slot = "val",
  var = "PointEst",
  notibble = FALSE
)
}
\arguments{
\item{reftab}{tibble with site and place}

\item{i}{Environment with indicators}

\item{props}{Character. Names of individual indicators}

\item{slot}{Character. Name of slot in indicator}

\item{var}{Character. Name of variable in slot}

\item{notibble}{Logical. Are the results contained in a tibble}
}
\description{
Make a tibble with values for a number of indicators per site and place
}
