% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monticarleer.R
\name{monticarleer}
\alias{monticarleer}
\title{monticarleer}
\usage{
monticarleer(prmlst, my.formula = "a * b ^ c", n = 100,
  df_out = FALSE, plot = TRUE, verbose = FALSE)
}
\arguments{
\item{prmlst}{List with items of the class distlist}

\item{n}{Integer. Number of samples to draw}

\item{df_out}{Logical. Return a dataframe}

\item{plot}{Logical. Plot or not}
}
\description{
Do a Monte Carlo simulation with a list of distribution proporties and a formula using them
}
