% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popProject.R
\name{popProjek}
\alias{popProjek}
\title{popProjek}
\usage{
popProjek(df = qol, var = "sol_energy_coal_ignition",
  place = "place", pop = data.frame(eMbalenhle = 35404, Lebohang =
  8908, eMzinoni = 10310, KwaDela = 982), latex = TRUE, xlsx = TRUE,
  dirr = tabdir, verbose = FALSE, debug = FALSE, c.loc = "bottom")
}
\arguments{
\item{df}{Data frame containing data to be used}

\item{var}{Character string that specifies the variable under consideration}

\item{pop}{Data frame containing populations}

\item{latex}{Logical that outputs latex table if TRUE}

\item{xlsx}{Logical that writes xlxs tables if TRUE}

\item{dirr}{Directory that output will be stored in}

\item{verbose}{Logical displaying message is TRUE}

\item{debug}{Logical Initializes debugging if TRUE}
}
\description{
Projects a sample population onto population
}
