% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dorp_brandstof.R
\name{resample_dorp_brandstof}
\alias{resample_dorp_brandstof}
\title{resample_dorp_brandstof}
\usage{
resample_dorp_brandstof(
  dta,
  placename = "place",
  unitname = "sol_energy_coal_consumptionwinter",
  formatname = "sol_energy_coal_format",
  pop,
  unitweights,
  notusename = "none",
  rep = 1000,
  ...
)
}
\arguments{
\item{dta}{data frame giving indication of proportions who use the fuel and format}

\item{placename}{Character. Default: "place"}

\item{unitname}{Character. Default: "sol_energy_coal_consumptionwinter"}

\item{formatname}{Character. Default: "sol_energy_coal_format"}

\item{pop}{dataframe}

\item{unitweights}{Named list with unit weights}

\item{notusename}{Character. Default: "none"}

\item{...}{Currently ignored}

\item{debug}{Logical}
}
\description{
Function to estimate township fuel using resampling. Used within boot_dorp_brandstof
}
