% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roep_param.R
\name{roep_param}
\alias{roep_param}
\title{roep_param}
\usage{
roep_param(l, x = NULL, n = NULL, verbose = FALSE)
}
\arguments{
\item{l}{distlist}

\item{x}{Numeric. Sobol sequence between 0 and 1 or NULL. If NULL, will be generated using sobol(n)}

\item{n}{Integer. Length of Sobol sequence or NULL. Will be used if x is NULL}

\item{verbose}{Logical. Talk back or not}
}
\description{
Generate n or length(x) values from the distribution described in distlist l
}
