% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skat_dorp_brandstof.R
\name{skat_dorp_brandstof}
\alias{skat_dorp_brandstof}
\title{skat_dorp_brandstof}
\usage{
skat_dorp_brandstof(pop, dta, placename = "place",
  unitname = "sol_energy_coal_consumptionwinter",
  formatname = "sol_energy_coal_format", notusename = "none",
  unitweights, som = FALSE, debug = FALSE)
}
\arguments{
\item{pop}{dataframe}

\item{dta}{data frame giving indication of proportions who use the fuel and format}

\item{placename}{Character. Default: "place"}

\item{unitname}{Character. Default: "sol_energy_coal_consumptionwinter"}

\item{formatname}{Character. Default: "sol_energy_coal_format"}

\item{notusename}{Character. Default: "none"}

\item{unitweights}{Named list with unit weights}

\item{som}{Logical}

\item{debug}{Logical}
}
\description{
Function to estimate township fuel. You need: a
population data frame, a data frame giving indication of proportions who use
the fuel and format, a consumption variable in unit counts, a dataframe with
format weights (1 if units are already standardised)
lapply this on a bootstrap sample to estimate uncertainty
}
