% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.indicator.R
\name{tibble.indicator.plek}
\alias{tibble.indicator.plek}
\title{tibble.indicator.plek}
\usage{
tibble.indicator.plek(x, nms = c("name", "description", "val", "unit"))
}
\arguments{
\item{x}{indicator}
}
\description{
Turn an indicator object into a tibble with site and town variable
}
